+++ av1decoder/include/decode.h
@@ -137,21 +137,21 @@ public:
 	int reconstruct(int plane, int x, int y, int txSz,BlockData *b_data,AV1DecodeContext *av1Ctx);
 
 
-	int inverseDCTArrayPermutation(int T[],int n);
-	int inverseDCT(int T[], int n, int r);
-	void inverseADSTInputArrayPermutation(int* T, int n);
-	void inverseADSTOutputArrayPermutation(int* T, int n);
-	void inverseADST4(int* T, int r);
-	void inverseADST8(int* T, int r);
-	void inverseADST16(int* T, int r);
-	void inverseADST(int T[],int n,int r);
-	void inverseWalshHadamardTransform(int* T, int shift) ;
-	void inverseIdentityTransform4(int* T);
-	void inverseIdentityTransform8(int* T);
-	void inverseIdentityTransform16(int* T);
-	void inverseIdentityTransform32(int* T);
-	void inverseIdentityTransform(int *T,int n);
-	void twoDInverseTransformBlock(int txSz,uint16_t **Residual,BlockData *b_data,AV1DecodeContext *av1Ctx) ;
+	int inverseDCTArrayPermutation(int16_t T[],int n);
+	int inverseDCT(int16_t T[], int n, int r);
+	void inverseADSTInputArrayPermutation(int16_t* T, int n);
+	void inverseADSTOutputArrayPermutation(int16_t* T, int n);
+	void inverseADST4(int16_t* T, int r);
+	void inverseADST8(int16_t* T, int r);
+	void inverseADST16(int16_t* T, int r);
+	void inverseADST(int16_t T[],int n,int r);
+	void inverseWalshHadamardTransform(int16_t* T, int shift) ;
+	void inverseIdentityTransform4(int16_t* T);
+	void inverseIdentityTransform8(int16_t* T);
+	void inverseIdentityTransform16(int16_t* T);
+	void inverseIdentityTransform32(int16_t* T);
+	void inverseIdentityTransform(int16_t *T,int n);
+	void twoDInverseTransformBlock(int txSz,int16_t **Residual,BlockData *b_data,AV1DecodeContext *av1Ctx) ;
 
 
 	void loopFilter(AV1DecodeContext *av1Ctx);
